Safari-609.1.11.1/LayoutTests/fast/encoding/charset-softbank-sjis-expected.txt

PASS document.characterSet is 'UTF-8'
PASS 'é' is 'é'
PASS successfullyParsed is true

TEST COMPLETE
